+++ src/csn.c
@@ -1,5 +1,7 @@
 #include "internal.h"
 
+int g_search_options;
+
 /* === STATIC FUNCTIONS === */
 /* write curl output to buffer that's ready to be used by tidy
  */
@@ -8,74 +10,6 @@ static uint write_cb(char *in, uint size, uint nmemb, TidyBuffer *out) {
     r = size * nmemb;
     tidyBufAppend(out, in, r);
     return r;
-}
-
-/* builds search url from options, currently lthe limit options is disable
- * but in the future, maybe we'll make more than one request to get some
- * further pages in order to get the number of results as user wanted
- */
-static char *build_search_url(const char *str, int options, int limit) {
-#ifdef ENABLE_DEBUG
-    _t_start = clock();
-#endif
-    buf_t *search_type = csn_buf_new(0);
-    buf_t *search_sort = csn_buf_new(0);
-    buf_t *search_cat  = csn_buf_new(0);
-
-    /* failed to do it the more rational way due to the limitation of
-     * C macro system
-     */
-#define TYPE_CASE(op, s) \
-    if (options & op) { \
-        csn_buf_write(search_type, s); \
-    }
-
-#define SORT_CASE(op, s) \
-    if (options & op) { \
-        csn_buf_write(search_sort, s); \
-    }
-
-#define CAT_CASE(op, s) \
-    if (options & op) { \
-        csn_buf_write(search_cat, s); \
-    }
-
-    TYPE_CASE(SEARCH_ARTIST, CSN_S_SEARCH_ARTIST);
-    TYPE_CASE(SEARCH_SONG, CSN_S_SEARCH_SONG);
-    TYPE_CASE(SEARCH_COMPOSER, CSN_S_SEARCH_COMPOSER);
-    TYPE_CASE(SEARCH_ALBUM, CSN_S_SEARCH_ALBUM);
-    TYPE_CASE(SEARCH_LYRICS, CSN_S_SEARCH_LYRICS);
-
-    SORT_CASE(SEARCH_SORT_MOST_LOVED, CSN_S_SEARCH_SORT_MOST_LOVED);
-    SORT_CASE(SEARCH_SORT_BEST_QUALITY, CSN_S_SEARCH_SORT_BEST_QUALITY);
-    SORT_CASE(SEARCH_SORT_LATEST, CSN_S_SEARCH_SORT_LATEST);
-
-    CAT_CASE(SEARCH_CATEGORY_MUSIC, CSN_S_SEARCH_CATEGORY_MUSIC);
-    CAT_CASE(SEARCH_CATEGORY_BEAT, CSN_S_SEARCH_CATEGORY_BEAT);
-    CAT_CASE(SEARCH_CATEGORY_VIDEO, CSN_S_SEARCH_CATEGORY_VIDEO);
-
-#undef TYPE_CASE
-#undef SORT_CASE
-#undef CAT_CASE
-
-    logf("search options:\ntype: %s\nsort: %s\ncat: %s\n", search_type->str, search_sort->str, search_cat->str);
-
-    char temp[1024]; //TODO: fix this
-    const char *search_fmt = CSN_SEARCH_URL"?s=%s&mode=%s&order=%s&cat=%s";
-
-    sprintf(temp, search_fmt, str, search_type->str, search_sort->str, search_cat->str);
-    logf("%s\n", temp);
-
-/* free stuff
- */
-    csn_buf_free(search_type);
-    csn_buf_free(search_sort);
-    csn_buf_free(search_cat);
-#ifdef ENABLE_DEBUG
-    _t_end = clock();
-    logf("Took %ld to complete\n", _t_end - _t_start);
-#endif
-    return strdup(temp);
 }
 
 /* === API PUBLIC FUNCTIONS === */
@@ -139,7 +73,12 @@ csn_result_t *csn_search(csn_ctx_t *ctx, const char *str, int options, int limit
             if (ctx->err >= 0) {
                 ctx->err = tidyRunDiagnostics(ctx->tdoc); /* load tidy error buffer */
                 if (ctx->err >= 0) {
-                    return parse_search_result(ctx->tdoc); /* walk the tree */
+                    if (g_search_options & SEARCH_ALBUM) {
+                        return NULL; // TODO:
+                    }
+                    else {
+                        return parse_song_search_result(ctx->tdoc); /* walk the tree */
+                    }
                 }
             }
         }
@@ -168,6 +107,48 @@ csn_album_info_t *csn_fetch_album_info(csn_ctx_t *ctx, csn_album_t *a) {
 
 csn_album_info_t *csn_fetch_album_info_url(csn_ctx_t *ctx, const char *url) {
     return NULL;
+}
+
+/* function to create csn objects
+ */
+csn_download_t *csn_download_new() {
+    csn_download_t *dl = xcalloc(sizeof(csn_download_t));
+    return dl;
+}
+
+csn_song_t *csn_song_new(int type) {
+    csn_song_t *s = xcalloc(sizeof(csn_song_t));
+    s->type = type;
+
+    return s;
+}
+
+csn_album_t *csn_album_new() {
+    csn_album_t *a = xcalloc(sizeof(csn_album_t));
+    return a;
+}
+
+csn_song_info_t *csn_song_info_new() {
+    csn_song_info_t *si = xcalloc(sizeof(csn_song_info_t));
+    return si;
+}
+
+csn_album_info_t *csn_album_info_new() {
+    csn_album_info_t *ai = xcalloc(sizeof(csn_album_info_t));
+    return ai;
+}
+
+csn_result_t *csn_result_new(bool is_song) {
+    csn_result_t *r = xcalloc(sizeof(csn_result_t));
+
+    r->is_song = is_song;
+    if (r->is_song) {
+        r->song = csn_song_new(CSN_TYPE_SONG);
+    }
+    else {
+        r->album = csn_album_new();
+    }
+    return r;
 }
 
 /* functions to free memory
